requirements.txt
-e git+ssh://git@github.com/edgemindstudio/gcs-core.git@d28176f999b7381432915aea3e2f019bdca25eaa#egg=gcs_core&subdirectory=../../gcs-core
absl-py==2.3.1
gradio>=4.0
astunparse==1.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
markdown-it-py==4.0.0
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.3.3
optree==0.17.0
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==6.32.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
tensorboard-data-server==0.7.2
tensorboard==2.20.0
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
